fix linter

+++ board/drivers/gpio.h
@@ -69,7 +69,7 @@ int get_gpio_input(GPIO_TypeDef *GPIO, unsigned int pin) {
 }
 
 void gpio_set_all_output(const gpio_t *pins, uint8_t num_pins, bool enabled) {
-  for(int i = 0; i < num_pins; i++) {
+  for (uint8_t i = 0; i < num_pins; i++) {
     set_gpio_output(pins[i].bank, pins[i].pin, enabled);
   }
 }
